Validations and Custom Exceptions

+++ src/main/java/com/devsuperior/desafio/services/ClientService.java
@@ -3,8 +3,11 @@ package com.devsuperior.desafio.services;
 import com.devsuperior.desafio.dtos.ClientDTO;
 import com.devsuperior.desafio.entities.Client;
 import com.devsuperior.desafio.repositories.ClientRepository;
+import com.devsuperior.desafio.services.exceptions.DatabaseException;
 import com.devsuperior.desafio.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -19,7 +22,7 @@ public class ClientService {
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id){
         Client entity = repository.findById(id).orElseThrow(
-                () -> new ResourceNotFoundException("Cliente não encontrado com o ID:" + id));
+                () -> new ResourceNotFoundException("Cliente não encontrado com o ID: " + id));
         return new ClientDTO(entity);
     }
 
@@ -39,15 +42,29 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto){
-        Client entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
+        try {
+            Client entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
+        }
+        catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Cliente não encontrado com o ID: " + id);
+        }
     }
 
     @Transactional
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Cliente não encontrado com o ID: " + id);
+        }
+
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Cliente não pode ser deletado. Entidade em conflito.");
+        }
     }
 
     private void copyDtoToEntity(ClientDTO dto, Client entity) {
